exp10_loop_invariant_code_motion is updated
